apps/api/pyproject.toml
[project]
name = "ai-copilot-api"
version = "0.1.0"
description = "AI Copilot HTTP API (MVP through Phase 9 batch adequacy)."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic>=1.14.0",
    "apscheduler>=3.10.4",
    "argon2-cffi>=23.1.0",
    "boto3>=1.35.0",
    "email-validator>=2.2.0",
    "openpyxl>=3.1.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    "psycopg[binary]>=3.2.0",
    "pyjwt>=2.10.0",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
    "pypdf>=6.10.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
